+++ pajbot/bot.py
@@ -638,39 +638,103 @@ class Bot:
             return False
         return self.thread_locals.moderation_actions is not None
 
-    def _ban(self, login: str, reason: Optional[str] = None) -> None:
-        message = f"/ban {login}"
-        if reason is not None:
-            message += f" {reason}"
-        self.privmsg(message)
+    def _ban(self, user_id: str, reason: Optional[str] = None) -> None:
+        try:
+            self.twitch_helix_api.ban_user(self.streamer.id, self.bot_user.id, self.bot_token_manager, user_id, reason)
+        except HTTPError as e:
+            if e.response.status_code == 401:
+                log.error(f"Failed to ban user with id {user_id}, unauthorized: {e} - {e.response.text}")
+            else:
+                log.error(f"Failed to ban user with id {user_id}: {e} - {e.response.text}")
 
     def ban(self, user: User, reason: Optional[str] = None) -> None:
-        self.ban_login(user.login, reason)
+        self.ban_id(user.id, reason)
 
-    def ban_login(self, login: str, reason=None) -> None:
+    def ban_id(self, user_id: str, reason: Optional[str] = None) -> None:
+        self._ban(user_id, reason)
+
+    def ban_login(self, login: str, reason: Optional[str] = None) -> None:
+        user_id = self.twitch_helix_api.get_user_id(login)
+        if user_id is None:
+            log.error(f"Attempted to ban user with login {login}, but no such user was found")
+            return
+
         if self._has_moderation_actions():
             self.thread_locals.moderation_actions.add(login, Ban(reason))
         else:
             self.timeout_login(login, 30, reason, once=True)
-            self.execute_delayed(1, self._ban, login, reason)
+            self.execute_delayed(1, self.ban_id, user_id, reason)
+
+    def _unban(self, user_id: str) -> None:
+        try:
+            self.twitch_helix_api.unban_user(self.streamer.id, self.bot_user.id, user_id, self.bot_token_manager)
+        except HTTPError as e:
+            if e.response.status_code == 401:
+                log.error(f"Failed to unban user with id {user_id}, unauthorized: {e} - {e.response.text}")
+            else:
+                log.error(f"Failed to unban user with id {user_id}: {e} - {e.response.text}")
 
     def unban(self, user: User) -> None:
-        self.unban_login(user.login)
+        self.unban_id(user.id)
+
+    def unban_id(self, user_id: str) -> None:
+        self._unban(user_id)
 
     def unban_login(self, login: str) -> None:
+        user_id = self.twitch_helix_api.get_user_id(login)
+        if user_id is None:
+            log.error(f"Attempted to unban user with login {login}, but no such user was found")
+            return
+
         if self._has_moderation_actions():
             self.thread_locals.moderation_actions.add(login, Unban())
         else:
-            self.privmsg(f"/unban {login}")
+            self.unban_id(user_id)
+
+    def _untimeout(self, user_id: str) -> None:
+        try:
+            ban_data = self.twitch_helix_api.get_banned_user(
+                self.streamer.id, self.streamer_access_token_manager, user_id
+            )
+            if ban_data is None:
+                log.warning(f"User with ID {user_id} is already not banned or timed-out.")
+                return
+            # As per the Helix docs, expires_at will return empty if the user is permabanned.
+            if not ban_data.expires_at:
+                log.error(f"User with ID {user_id} is currently banned! Will not untimeout.")
+                return
+        except HTTPError as e:
+            if e.response.status_code == 401:
+                log.error(f"Failed to get banned user status with id {user_id}, unauthorized: {e} - {e.response.text}")
+                return
+            else:
+                log.error(f"Failed to get banned user status with id {user_id}: {e} - {e.response.text}")
+                return
+
+        try:
+            self.twitch_helix_api.unban_user(self.streamer.id, self.bot_user.id, user_id, self.bot_token_manager)
+        except HTTPError as e:
+            if e.response.status_code == 401:
+                log.error(f"Failed to untimeout user with id {user_id}, unauthorized: {e} - {e.response.text}")
+            else:
+                log.error(f"Failed to untimeout user with id {user_id}: {e} - {e.response.text}")
 
     def untimeout(self, user: User) -> None:
-        self.untimeout_login(user.login)
+        self.untimeout_id(user.id)
+
+    def untimeout_id(self, user_id: str) -> None:
+        self._untimeout(user_id)
 
     def untimeout_login(self, login: str) -> None:
+        user_id = self.twitch_helix_api.get_user_id(login)
+        if user_id is None:
+            log.error(f"Attempted to untimeout user with login {login}, but no such user was found")
+            return
+
         if self._has_moderation_actions():
             self.thread_locals.moderation_actions.add(login, Untimeout())
         else:
-            self.privmsg(f"/untimeout {login}")
+            self.untimeout_id(user_id)
 
     def _timeout(self, login: str, duration: int, reason: Optional[str] = None) -> None:
         message = f"/timeout {login} {duration}"
@@ -1018,7 +1082,7 @@ class Bot:
 
         if self.streamer == "forsen":
             if "zonothene" in login:
-                self._ban(login)
+                self._ban(id)
                 return True
 
             raw_m = event.arguments[0].lower()
@@ -1112,7 +1176,7 @@ class Bot:
 
     def commit_all(self) -> None:
         for key, manager in self.commitable.items():
-            manager.commit()
+            manager.commit()  # type: ignore[attr-defined]
 
         HandlerManager.trigger("on_commit", stop_on_false=False)