+++ pajbot/modules/ascii.py
@@ -7,7 +7,6 @@ log = logging.getLogger(__name__)
 
 
 class AsciiProtectionModule(BaseModule):
-
     ID = __name__.split(".")[-1]
     NAME = "ASCII Protection"
     DESCRIPTION = "Times out users who post messages that contain too many ASCII characters."