+++ CHANGELOG.md
@@ -8,9 +8,24 @@ For your bot to work after these changes, the bot owner **must** re-authenticate
 Remember to bring your dependencies up to date with `./scripts/venvinstall.sh` when updating to this version!
 
 - Breaking: Migrated Follower Only function from IRC to Helix. (#2175)
+- Breaking: Migrated untimeout function from IRC to Helix. (#2223)
+- Breaking: Migrated unbanning function from IRC to Helix. (#2222)
+- Breaking: Migrated banning function from IRC to Helix. (#2213)
+- Breaking: Migrated VIP Refresh module from IRC to Helix. (#2188)
+- Breaking: Migrated Moderator Refresh module from IRC to Helix. (#2186, #2202)
+- Breaking: Migrated Sub Mode function from IRC to Helix. (#2185)
+- Breaking: Migrated Slow Mode function from IRC to Helix. (#2176)
+- Breaking: Migrated emote only function from IRC to Helix. (#2178)
+- Breaking: Migrated Unique Chat function from IRC to Helix. (#2177)
 - Breaking: Migrated Announce from IRC to Helix. (#2141)
 - Breaking: Migrated Delete moderation action from IRC to Helix. (#2173)
+- Bugfix: Exclude deleted accounts from twitch subscribers list. (#2292)
+- Bugfix: Fix bot not correctly tracking online/offline state of stream due to TwitchGame not being deserialized properly. (#2243)
+- Minor: Updated `Wide Emote Limit` module to account for wide BTTV emotes (##2272)
 - Minor: Migrated LastFM module to the `reply` response type. (#2118, #2128)
+- Minor: Increased efficiency and speed of subscriber status refresh. (#2203)
+- Minor: Install documentation now recommends the use of limited-scope CloudFlare API tokens. (#2201)
+- Dev: Migrated to 7TV's new REST API. (#2268)
 
 ## v1.62